+++ libs/core/src/thds/core/log/__init__.py
@@ -24,6 +24,7 @@ logger.info("testing 5")
 ```
 """
 
+from . import env  # noqa: F401
 from .basic_config import DuplicateFilter, set_logger_to_console_level  # noqa: F401
 from .kw_formatter import ThdsCompactFormatter  # noqa: F401
 from .kw_logger import KwLogger, auto, getLogger, logger_context, make_th_formatters_safe  # noqa: F401